api/requirements.txt
fastapi
uvicorn[standard]
PyJWT
cryptography
supabase
httpx
